requirements.txt
# requirements.txt
pytest
pytest-asyncio
playwright
pytest-html
allure-pytest
pytest-xdist
pytest-rerunfailures